[cake] Initialize default target to interval / 16
Codel theory (see the codel RFC) argues that target should be in the
range of 5 to 10 percent of interval, with smaller targets sacrificing
slightly more bandwidth for decent latency under load. For DOCSIS 6.2 ms
is claimed to be slightly better than the previous default of 5ms.
The cake module or tc already changed the default to 100/16 ms, so this
change makes sure cake defaults to a target og 6.2ms even if no interval
parameter is passed from tc to sch_cake. Needs testing...

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>