Remove peel threshold and always peel GSO packets
How we calculate this has always been problematic. We are aiming
for the lowest possible latency here and *most* of the benefit
of GRO and TSO is in saving on routing cache lookups. xmit_more
will take care of handing off very small packets to the hardware
(hopefully)

furthermore, I think that the peeling routine's use
of
			q->buffer_used      += segs->truesize;

was not correct if we only optionally peeled.

Peel always.