When there is no rate set default quantum to 1514 not 300
The mechanics of the quantum parameter are tricky and hard to describe
but the effects on cpu when at 300 are you end up looping 5.x times
to actually deliver big packets.

When rate is unknown we might as well assume a full packet.

Arguably when rate is greater than a gbit, (possibly lower,
or dependent on the number of flows active) we should use an even
larger quantum that 1514.

(sch_fq uses a 3028 quantum, which releases two packets which in turn
release an ack from the other side - which is a good thing)