Sort out remaining subtleties in triple-isolation logic.
In the flow dissector, the overall hash function can't be run until the flow_keys struct is no longer needed as a spare copy of the src and dst addresses, because it can re-order the src and dst addresses for canonicalisation.  Forgetting this caused src and dst address to be treated as identical (losing one of them) about half the time.

It's also necessary to reset the count of host-deferred flows when a flow is deferred due to being in *flow* deficit rather than *host* deficit.  Otherwise they get double-counted if the flow quantum is less than MTU (which it often is).