rearrange diffserv3/4 tin orders
so diffserv3 & 4 are the only modes that use an effective tin order of 1
0 2 (3) which means that dequeue has to look into a table of tins to
work out in what order it should be doing them.  Why not re-arrange the
order in the tin data thus negating the need for this 'double lookup' at
dequeue.

For stats purposes we retain the tin order info so we can present tins
in the order that tc expects otherwise we'd have to change tc as well.