Add TCP ACK filtering
This is beneficial to the upstream of highly asymmetrical links, where
TCP ACKs alone are often enough to saturate the link. ACKs are
cumulative, so it is desirable to queue only the most recent ACK for a
connection at any point in time. This scheme is commonly implemented
within DOCSIS modems.

ACKs with duplicate acknowledgement sequence numbers (DupACKs) should
not be filtered as this interferes with fast retransmit. SACKs should
not be filtered either (without special handling), this patch relies on
the same property of DupACKs that they will have duplicate ack sequence
numbers, and does not implement any special handling of TCP options.
Special handling of SACKs should also be possible in the future.

This routine is triggered on the enqueue of a TCP packet (containing
data or a 'pure' ACK), and scans its corresponding flow queue for 'pure'
ACKs that may be filtered if they are stale and redundant.

Overhead is minimised in the case of an upload (where there may be a
deeper flow queue without any 'pure' ACKs to thin), by saving the point
in the queue that is known to be ACK-free. UDP and other protocols will
not trigger the routine.

ECE/CWR/NS can be safely ignored in ACKs, however 'pure' ACKs with any
other flags set will not be filtered to avoid future breakage.

As ACKs are anti-responsive, filtering only makes sense for egress.

An admittedly somewhat unscientific test indicates huge gains when
tested on a LAN. Bidirectional netcat + pv transfer with upstream shaped
by cake to 300kbps.

without ack filtering:
4MiB/s or 36MiB/s (depending on whether BLUE happens to kick in)
upload 18KiB/s

with ack filtering:
download 71MiB/s
upload 30KiB/s

This is a performance win/win, importantly it is consistent unlike BLUE.
A less aggressive variant that does not drop ACKs from the head of the
queue was also tested and yielded ~75MiB/s down 23KiB/s up. It is
considered that the somewhat more aggressive recovery of bandwidth from
ACKs towards data flows is desirable in most cases with no evident
drawbacks.

Signed-off-by: Ryan Mounce <ryan@mounce.com.au>