ack_filter: make less aggressive by default
Aims to mitigate some situations where the naive implementation could be
overzealous.
Only filter if there are more than two ACKs in the flow queue.

TCP increases the window size for every ACK received during slow-start.
Keeping up to two ACKs in the queue avoids buffer underrun during this
critical phase of the connection.

Another potential issue comes from reducing 'redundancy' in the stream
of ACKs. This isn't an issue in long running transfers, however if the
second-to-last ACK for a burst e.g. HTTP GET is filtered and the last
ACK is lost elsewhere this could result in a large retransmission. By
never filtering the second-to-last ACK in a burst the maximum redundant
retransmission due to a lost ACK is limited to 2x MSS in most cases.

Signed-off-by: Ryan Mounce <ryan@mounce.com.au>