sch_cake.c: add MODULE_VERSION to baseline and track future ABI changes
Prior to upstreaming, CAKE has been a moving target in terms of its ABI,
with frequent updates to keywords and functionality. After upstreaming,
changes are expected to slow but still continue until CAKE stabilizes.

To provide users with an accessible indicator of CAKE functionality going
forward, include an ABI version both as a static string and via the
MODULE_VERSION macro, with a common version string format ("version=").

As with the existing inclusion of the git revision, using both a module
macro and static string should be more robust in the face of aggressive
module stripping such as in LEDE/OpenWRT. This change also helps avoid
encouraging other, more fragile approaches to CAKE qdisc introspection,
such as ad-hoc parsing of related system man pages or help text.

Also clarify that the comment "version 5" at top-of-file refers to the
statistics structure revision. Perhaps better to just remove this?

Signed-off-by: Tony Ambardar <guidosarducci@users.noreply.github.com>