Do the full header length calculation for GSO overhead compensation
The previous version where we just divided the length by the number of
segments fails if the last segment is shorter. Borrow the header
calculation from qdisc_pkt_len_init() to get the right length.

We account the last segment as the full size even if it is shorter, though,
to avoid having to duplicate the whole size calculation for that last
segment.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>