ack_filter: Add proper handling of SACKs
Previously when presented with multiple SACKs, a newer SACK acknowleding
more data could be filtered.

This patch adds a check to ensure that the blocks refered to in a SACK
considered for filtering are a strict subset of a newer SACK.
In the common case of packet loss on a high BDP flow, this will allow
the ACK filter to remain effective.

Signed-off-by: Ryan Mounce <ryan@mounce.com.au>