Obey tin_order for tc filter classifiers
Everywhere we expose the tins to userspace, we use the tin_order to reorder
them, so the bulk tin is logically the first one, even though the internal
representation is different. This mapping is missing when selecting a tin
based on the result of a tc filter classifier. Fix this by applying the
order lookup to the filter result.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>