Reinstate TC filter flow mapping override and expand it to hosts as well
This re-enables the ability to override the flow hashing from tc filter,
but makes sure cake_hash() is always called so the bookkeeping code is run
correctly.

In addition, the major number of the class ID can now be set to override
the host mapping in host isolation mode. If both host and flow are
overridden (or if the respective modes are disabled), flow dissection and
hashing will be skipped entirely; otherwise, the hashing will be kept for
the portions that are not set by the filter.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>