compat: Don't lock root qdisc when dumping stats on old kernels
The qdisc stats dumping was changed in 4.8 (by commit
edb09eb17ed89eaa82a52dd306beac93e292b485) to be lockless and moving the
responsibility for taking the qdisc root lock into the qdisc itself. Cake
inherited this from FQ-CoDel, but it leads to deadlocks on old kernels
where the core stats dumping code still takes the root qdisc lock.

This adds a conditional wrapper that is set in compat to hopefully fix this
deadlock.

Reported-by: Pete Heist <pete@heistp.net>
Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>