Make the dual modes fairer
Host fairness does not work well with TCP flows in dual-srchost and
dual-dsthost setup. The reason is that ACKs generated by TCP flows
are classified as sparse flows, and affect flow isolation from other
hosts. Fix this by calculating host_load based only on the bulk flows
a host generates. In a hash collision the host_bulk_flow_count values must
be decremented on the old hosts and incremented on the new ones *if* the
queue is in the bulk set.

Reported-by: Pete Heist <peteheist@gmail.com>
Signed-off-by: George Amanakis <gamanakis@gmail.com>