compat: Unlock qdisc root lock in case of failure when dumping stats.
Linux versions prior to 4.8 have a bug where the root qdisc lock is
not released in tc_fill_qdisc (in sch_api.c) in some failure cases.
This can cause a lockup when dumping cake stats from multiple
qdiscs, for example. Fix this by explicitly unlocking the qdisc
lock in the failure case.