Remove ewma from some delay calculations
The smoothing here was excessive. Due to not wanting to change the
API, and since we are sampling at 10ms elsewhere, this change
clears the existing delay statistic when polled, which gives a
much more accurate picture as to what is really going on "down there".

We may end up using these vars for some other things now.