sch_cake: Convert the invsqrt table to const
I have no idea why we did not convert this to const long ago,
perhaps because I had little faith that the invsqrt method we
used was correct, or perhaps because I wanted to fiddle with the
curve later.

This just takes the formerly generated invsqrt table and
slams in the constants. It might be a bit off from what
proper floating point would generate, will check.